added viewmodel and viewmodelfactory (not working for now)
