+++ app/src/main/java/com/example/f1calendarOP/RaceListFragment.kt
@@ -6,6 +6,8 @@ import android.view.LayoutInflater
 import androidx.fragment.app.Fragment
 import android.view.View
 import android.view.ViewGroup
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.DefaultItemAnimator
@@ -18,6 +20,7 @@ const val TAG = "RaceListFragment"
 
 class RaceListFragment : Fragment(R.layout.fragment_race_list) {
 
+    private lateinit var viewModel: RaceListViewModel
     private val raceListAdapter: RaceListAdapter by lazy {
         RaceListAdapter { race -> onClickHelper(race) }
     }
@@ -40,28 +43,33 @@ class RaceListFragment : Fragment(R.layout.fragment_race_list) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        lifecycleScope.launchWhenCreated {
-            val response = try {
-                RetrofitInstance.api.getRaceInfo()
-            } catch (e: IOException) {
-                Log.e(TAG, e.message.toString())
-                return@launchWhenCreated
-            } catch (e: HttpException) {
-                Log.e(TAG, e.message.toString())
-                return@launchWhenCreated
-            }
-            if (response != null) {
-                val responseRaceList: List<Race> = response.mrData.raceTable.races
-                val raceFunctions = RaceFunctions()
-                for (item in responseRaceList) {
-                    item.flagImage = raceFunctions.getFlagByCountry(item.circuit.location.country)
-                    item.weekendDate = raceFunctions.getWeekendDate(item)
-                }
-                raceListAdapter.submitList(responseRaceList)
-            } else {
-                Log.e(TAG, "Response not successful")
-            }
+//        lifecycleScope.launchWhenCreated {
+//            val response = try {
+//                RetrofitInstance.api.getRaceInfo()
+//            } catch (e: IOException) {
+//                Log.e(TAG, e.message.toString())
+//                return@launchWhenCreated
+//            } catch (e: HttpException) {
+//                Log.e(TAG, e.message.toString())
+//                return@launchWhenCreated
+//            }
+//            if (response != null) {
+//                val responseRaceList: List<Race> = response.mrData.raceTable.races
+//                val raceFunctions = RaceFunctions()
+//                for (item in responseRaceList) {
+//                    item.flagImage = raceFunctions.getFlagByCountry(item.circuit.location.country)
+//                    item.weekendDate = raceFunctions.getWeekendDate(item)
+//                }
+//                raceListAdapter.submitList(responseRaceList)
+//            } else {
+//                Log.e(TAG, "Response not successful")
+//            }
+//        }
+
+        viewModel.raceList.observe(viewLifecycleOwner) {
+            raceListAdapter.submitList(it)
         }
+        viewModel.getAllRaces()
     }
 
     private fun onClickHelper(race: Race) {
